lambda/api/requirements.txt
fastapi
uvicorn
setuptools
python-multipart
psycopg2-binary
boto3
python-jose[cryptography]
PyMuPDF
Pillow
opencv-python-headless
requests
python-magic
